move zmq.KernelManagers into IPython.kernel
