+++ IPython/kernel/blockingkernelmanager.py
